LPS-54975 Move provider classes to portlet dir
