+++ modules/apps/asset/asset-publisher-web/src/com/liferay/asset/publisher/web/portlet/AssetPublisherPortletProvider.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.asset.publisher.web.provider;
+package com.liferay.asset.publisher.web.portlet;
 
 import com.liferay.asset.publisher.web.constants.AssetPublisherPortletKeys;
 import com.liferay.asset.publisher.web.util.AssetPublisherUtil;
-import com.liferay.portal.kernel.provider.AddPortletProvider;
-import com.liferay.portal.kernel.provider.PortletProvider;
-import com.liferay.portal.kernel.provider.ViewPortletProvider;
+import com.liferay.portal.kernel.portlet.AddPortletProvider;
+import com.liferay.portal.kernel.portlet.PortletProvider;
+import com.liferay.portal.kernel.portlet.ViewPortletProvider;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portlet.asset.model.AssetEntry;
 import com.liferay.portlet.asset.service.AssetEntryLocalServiceUtil;
