+++ modules/apps/journal/journal-content-web/src/META-INF/resources/init.jsp
@@ -29,8 +29,8 @@
 <%@ page import="com.liferay.journal.content.web.context.JournalContentDisplayContext" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.portlet.LiferayWindowState" %><%@
-page import="com.liferay.portal.kernel.provider.PortletProvider" %><%@
-page import="com.liferay.portal.kernel.provider.PortletProviderUtil" %><%@
+page import="com.liferay.portal.kernel.portlet.PortletProvider" %><%@
+page import="com.liferay.portal.kernel.portlet.PortletProviderUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
 page import="com.liferay.portal.kernel.util.HtmlUtil" %><%@
 page import="com.liferay.portal.kernel.util.ParamUtil" %><%@
