+++ modules/apps/journal/journal-content-web/src/com/liferay/journal/content/web/portlet/JournalContentDisplayInformationProvider.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.journal.content.web.provider;
+package com.liferay.journal.content.web.portlet;
 
 import com.liferay.journal.content.web.constants.JournalContentPortletKeys;
-import com.liferay.portal.kernel.provider.DisplayInformationProvider;
+import com.liferay.portal.kernel.portlet.DisplayInformationProvider;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portlet.journal.model.JournalArticle;
