+++ portal-impl/src/com/liferay/portlet/journal/service/impl/JournalContentSearchLocalServiceImpl.java
@@ -18,7 +18,7 @@ import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.provider.DisplayInformationProvider;
+import com.liferay.portal.kernel.portlet.DisplayInformationProvider;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.model.LayoutTypePortlet;
