+++ portal-service/src/com/liferay/portal/kernel/portlet/ViewPortletProvider.java
@@ -12,28 +12,10 @@
  * details.
  */
 
-package com.liferay.site.navigation.language.web.provider;
-
-import com.liferay.portal.kernel.provider.ViewPortletProvider;
-import com.liferay.site.navigation.language.web.constants.LanguagePortletKeys;
-
-import org.osgi.service.component.annotations.Component;
+package com.liferay.portal.kernel.portlet;
 
 /**
  * @author Eudaldo Alonso
  */
-@Component(
-	immediate = true,
-	property = {
-		"model.class.name=com.liferay.portal.kernel.servlet.taglib.ui.LanguageEntry"
-	},
-	service = ViewPortletProvider.class
-)
-public class LanguageEntryViewPortletProvider implements ViewPortletProvider {
-
-	@Override
-	public String getPortletId() {
-		return LanguagePortletKeys.LANGUAGE;
-	}
-
+public interface ViewPortletProvider extends PortletProvider {
 }
