+++ util-taglib/src/com/liferay/taglib/util/VelocityTaglib.java
@@ -14,7 +14,7 @@
 
 package com.liferay.taglib.util;
 
-import com.liferay.portal.kernel.provider.PortletProvider;
+import com.liferay.portal.kernel.portlet.PortletProvider;
 import com.liferay.portal.kernel.servlet.taglib.ui.Menu;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.model.Portlet;
